+++ .gitignore
@@ -1,3 +1,3 @@
 .vscode
 .DS_Store
- */__debug_bin
+.idea/
